pyproject.toml
[project]
name = "tool_shop"
version = "0.1.0"
description = ""
authors = [
    {name = "Yana Zaitseva", email = "zaitseva.yg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry.dependencies]
python = ">=3.13,<4.0"

# CORE
pytest = ">=8,<10"
pytest-playwright = "^0.7.2"
playwright = "^1.47.0"
allure-pytest = "^2.13.3"

# API
requests = "^2.31.0"
jsonschema = "4.26.0"
python-dotenv = "^1.2.1"
pytest-asyncio = "^1.3.0"
schemathesis = "^4.9.3"
openapi-core = "^0.22.0"
openapi-spec-validator = "^0.7.2"



[tool.poetry]
packages = [{include = "tool_shop", from = "src"}]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Playwright
[tool.playwright]
chromium = true
firefox = false
webkit = false

[dependency-groups]
dev = [
    "pytest-html (>=4.2.0,<5.0.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)"
]
